Added inline "if" statements
Sentences starting with `if` can be used to control the flow. An `if` sentence
takes on of the following forms:

    if <condition>, <true>

    if <condition>, <true>, otherwise <false>

Where `<condition>` is a simple comparison between two variables or values. Some
examples are:

    name = "Bob"
    counter > 10

All supported operators are:

- `=` - Equal.
- `!=` - Not equal.
- `>` - Greater than.
- `>=` - Greater than or equal.
- `<` - Less than.
- `<=` - Less than or equal.

Values can only be compared when they are the same type. For example the
following is not allowed, and will return an error:

    "123" = 123