Added special "_" (blackhole) variable
The blackhole variable is a single underscore (`_`). It can be used as a placeholder when the value can be ignored. For example:

    divide 1.23 by 7.89 into _

If `_` is used in a place where the value would be read it will return a zero value (the same as the default value for that type).

The following lines would function the same. However, you should not rely on the blackhole variable as a readable value and it may be made illegal in the future:

    add _ and 7.89 into result
    add 0 and 7.89 into result

Fixes #32