Added Questions
A question is a special type of function that is defined with a `?` instead of a
`:`:

    it is ok?
        yes

A question is answered with the `yes` or `no` sentences. Once a question is
answered it will return immediately.

If a question is not explicitly answered by the end, it's assumed to be `no`.

Questions can be asked in conditionals:

    start:
        if it is ok, display "All good!"

Questions can also take arguments in the same way that functions do:

    start:
        declare x is number

        set x to 123
        if x is over 100, display "It's over 100", otherwise display "Not yet"

    the-number is over threshold (the-number is number, threshold is number)?
        if the-number > threshold, yes

Fixes #21